Add folds data using RNAfold.
The first number in the text file after a sequence is the free energy.
The second number in the text file after that is the measure for the
"well-definedness of the minimum free energy structure".

The first number is the one we should really use, the second is
optional? Must verify with prof.